+++ Tangra.Core/Tangra.Image.h
@@ -9,7 +9,7 @@
 #define MAX_CONVOLUTION_WIDTH 1280
 #define MAX_CONVOLUTION_HEIGHT 1024
 
-unsigned long RESULT_BUFFER[MAX_CONVOLUTION_WIDTH * MAX_CONVOLUTION_HEIGHT];
+unsigned int RESULT_BUFFER[MAX_CONVOLUTION_WIDTH * MAX_CONVOLUTION_HEIGHT];
 
 const double GAUSSIAN_BLUR_MATRIX[] = { 
 			1 / 16.0, 1 / 8.0, 1 / 16.0 , 
@@ -26,9 +26,9 @@ const double DENOISE_MATRIX[] = {
 			1/9.0, 1/9.0, 1/9.0, 
             1/9.0, 1/9.0, 1/9.0 };
 			
-void Convolution_GaussianBlur(unsigned long* pixels, long bpp, long width, long height);
-void Convolution_Sharpen(unsigned long* pixels, long bpp, long width, long height, unsigned long* average);
-void Convolution_Denoise(unsigned long* pixels, long bpp, long width, long height, unsigned long* average, bool cutEdges);
+void Convolution_GaussianBlur(unsigned int* pixels, int bpp, int width, int height);
+void Convolution_Sharpen(unsigned int* pixels, int bpp, int width, int height, unsigned int* average);
+void Convolution_Denoise(unsigned int* pixels, int bpp, int width, int height, unsigned int* average, bool cutEdges);
 
 /* Make sure functions are exported with C linkage under C++ compilers. */
 #ifdef __cplusplus
@@ -36,10 +36,10 @@ extern "C"
 {
 #endif
 
-DLL_PUBLIC HRESULT Convolution(unsigned long* data, long bpp, long width, long height, const double* convMatrix, bool cutEdges, bool calculateAverage, unsigned long* average);
-DLL_PUBLIC HRESULT PrepareImageForOCR(unsigned long* pixels, long bpp, long width, long height);
-DLL_PUBLIC HRESULT PrepareImageForOCRSingleStep(unsigned long* pixels, long bpp, long width, long height, long stepNo, unsigned long* average);
-DLL_PUBLIC HRESULT LargeChunkDenoise(unsigned long* pixels, long width, long height, unsigned long onColour, unsigned long offColour);
+DLL_PUBLIC HRESULT Convolution(unsigned int* data, int bpp, int width, int height, const double* convMatrix, bool cutEdges, bool calculateAverage, unsigned int* average);
+DLL_PUBLIC HRESULT PrepareImageForOCR(unsigned int* pixels, int bpp, int width, int height);
+DLL_PUBLIC HRESULT PrepareImageForOCRSingleStep(unsigned int* pixels, int bpp, int width, int height, int stepNo, unsigned int* average);
+DLL_PUBLIC HRESULT LargeChunkDenoise(unsigned int* pixels, int width, int height, unsigned int onColour, unsigned int offColour);
 
 #ifdef __cplusplus
 } // __cplusplus defined.
